+++ tests/api/test_constants.py
@@ -2,7 +2,7 @@
 
 from types import SimpleNamespace
 
-from cosmology.api import CosmologyConstantsAPINamespace
+from cosmology.api import CosmologyConstantsNamespace
 
 ################################################################################
 # TESTS
@@ -12,26 +12,26 @@ from cosmology.api import CosmologyConstantsAPINamespace
 def test_noncompliant_namespace():
     """
     Test that a non-compliant namespace is not a
-    `cosmology.api.CosmologyConstantsAPINamespace`.
+    `cosmology.api.CosmologyConstantsNamespace`.
     """
     example_namespace = SimpleNamespace()
 
-    assert not isinstance(example_namespace, CosmologyConstantsAPINamespace)
+    assert not isinstance(example_namespace, CosmologyConstantsNamespace)
 
 
 def test_compliant_namespace():
     """
     Test that a compliant namespace is a
-    `cosmology.api.CosmologyConstantsAPINamespace`.
+    `cosmology.api.CosmologyConstantsNamespace`.
     """
     example_namespace = SimpleNamespace(G=1, c=2)
 
-    assert isinstance(example_namespace, CosmologyConstantsAPINamespace)
+    assert isinstance(example_namespace, CosmologyConstantsNamespace)
 
 
 def test_fixture(constants_ns):
     """
     Test that the ``constants_ns`` fixture is a
-    `cosmology.api.CosmologyConstantsAPINamespace`.
+    `cosmology.api.CosmologyConstantsNamespace`.
     """
-    assert isinstance(constants_ns, CosmologyConstantsAPINamespace)
+    assert isinstance(constants_ns, CosmologyConstantsNamespace)