+++ tests/api/test_namespace.py
@@ -2,7 +2,7 @@
 
 from types import SimpleNamespace
 
-from cosmology.api import CosmologyAPINamespace
+from cosmology.api import CosmologyNamespace
 
 ################################################################################
 # TESTS
@@ -12,27 +12,27 @@ from cosmology.api import CosmologyAPINamespace
 def test_noncompliant_namespace():
     """
     Test that a non-compliant namespace is not a
-    `cosmology.api.CosmologyAPINamespace`.
+    `cosmology.api.CosmologyNamespace`.
     """
     example_namespace = SimpleNamespace()
 
-    assert not isinstance(example_namespace, CosmologyAPINamespace)
+    assert not isinstance(example_namespace, CosmologyNamespace)
 
 
 def test_compliant_namespace():
     """
     Test that a compliant namespace is a
-    `cosmology.api.CosmologyAPINamespace`.
+    `cosmology.api.CosmologyNamespace`.
     """
     constants_namespace = SimpleNamespace(G=1)
     example_namespace = SimpleNamespace(constants=constants_namespace)
 
-    assert isinstance(example_namespace, CosmologyAPINamespace)
+    assert isinstance(example_namespace, CosmologyNamespace)
 
 
 def test_fixture(cosmology_ns):
     """
     Test that the ``cosmology_ns`` fixture is a
-    `cosmology.api.CosmologyAPINamespace`.
+    `cosmology.api.CosmologyNamespace`.
     """
-    assert isinstance(cosmology_ns, CosmologyAPINamespace)
+    assert isinstance(cosmology_ns, CosmologyNamespace)